Ajout de la console grace une modification sommaire du main.c; ajustement de la vitesse dans jeu.c

+++ structure.h
@@ -16,7 +16,7 @@ typedef enum {TIR_LASER, OBUS, RAYON_LASER} _arme;
 
 typedef struct {
     int poid;
-    int vitesse; // En pixel par seconde
+    double vitesse; // En pixel par seconde
     int acceleration; // En prixel par seconde*pow(-2)
     int vitesse_max; // En pixel par seconde
     _intelligence intelligence;
